+++ CHANGELOG.md
@@ -10,7 +10,7 @@ A breaking change will get clearly marked in this log.
 ## [v12.3.0](https://github.com/stellar/js-stellar-sdk/compare/v12.2.0...v12.3.0)
 
 ### Added
-- Add `getTransactions` to RPC server. ([#1037](https://github.com/stellar/js-stellar-sdk/pull/1037))
+- `rpc.Server` now has a `getTransactions`, which has the same response schema as `getTransactions` except with bundles of transactions ([#1037](https://github.com/stellar/js-stellar-sdk/pull/1037)).
 - `rpc.Server` now has a `getVersionInfo` method which reports version information of the RPC instance it is connected to ([#997](https://github.com/stellar/js-stellar-sdk/issues/997)):
 
 ```typescript